requirement.txt
streamlit>=1.32.0
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-legacy>=0.9.0
llama-index-llms-openai-like>=0.1.0
matplotlib>=3.7.0
numpy>=1.24.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.15.0